Changes to comply with Holberton's betty standar Separated functionality in functions to comply with Holberton's betty standar
