+++ _exec.c
@@ -6,11 +6,13 @@
  *
  * @command: Command to execute with execve
  * @cmd_arg: Arguments for command to execute
+ * @exit_c: keeps track of command execution exit code
+ * @cmd_count: keeps track of number of commands executed
  */
 void _exec(char *command, char **cmd_arg, char *exit_c, int *cmd_count)
 {
-	pid_t child_pid;
 	int wait_status;
+	pid_t child_pid;
 
 	child_pid = fork();
 	if (child_pid == -1)
@@ -18,6 +20,9 @@ void _exec(char *command, char **cmd_arg, char *exit_c, int *cmd_count)
 	if (child_pid == 0)
 	{
 		execve(command, cmd_arg, NULL);
+		free(command);
+		free(*cmd_arg);
+		free(exit_c);
 	} else
 	{
 		wait(&wait_status);
@@ -27,20 +32,5 @@ void _exec(char *command, char **cmd_arg, char *exit_c, int *cmd_count)
 			wait_status = WEXITSTATUS(wait_status);
 			_itoa(wait_status, exit_c);
 		}
-	}
-}
-
-void _itoa(int wait_status, char *s)
-{
-	int i, status;
-
-	status = wait_status;
-	for (i = 0; status / 10 != 0; i++)
-		status = status / 10;
-	s[i + 1] = '\0';
-	for (; i >= 0; i--)
-	{
-		s[i] = (wait_status % 10) + '0';
-		wait_status = wait_status / 10;
 	}
 }
