+++ _free_all.c
@@ -1,7 +1,7 @@
 #include "shs.h"
 
 /**
- * _free_all - free memory used by variables
+ * itty_free - free memory used by variables when ran in interactive mode
  *
  * @pinput: pointer to memory used to store the user input
  * @cmd_arg: pointer to memory used to store the command arguments
@@ -14,6 +14,16 @@ void itty_free(char *pinput, char **cmd_arg, char *command)
 	free(cmd_arg);
 }
 
+/**
+ * ntty_free - free memory used by variables when ran in not
+ * interactive mode.
+ *
+ * @pinput: pointer to memory used to store the user input
+ * @cmd_arg: pointer to memory used to store the command arguments
+ * @command: pointer to mermory used to store the command to exec
+ * @env: pointer to mermory used to store environment PATH values
+ * @exit_c: pointer to memory used to store exit codes
+ */
 void ntty_free(char *pinput, char **cmd_arg,
 	       char *command, char *env, char *exit_c)
 {
